relocate the function to recursively check for undefined vars
add another test
